requirements.txt
django==3.2
selenium==4.1.3
undetected-chromedriver==1.5.2
celery==5.2.4
redis==4.2.2
django-celery-beat==2.3.0
djangorestframework==3.13.1
django-cors-headers==3.13.0
django-dotenv==1.4.2
psycopg2==2.9.3